server/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# ── Auth ──────────────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ── Database ──────────────────────────────────────────────────────────────────
# psycopg2-binary==2.9.10    # PostgreSQL (prod) — install separately when deploying
# sqlite3 is built-in for dev

# ── Environment ───────────────────────────────────────────────────────────────
django-environ==0.11.2

# ── Utilities ─────────────────────────────────────────────────────────────────
# Pillow==11.1.0             # image fields — install separately if needed
django-filter==24.3          # queryset filtering
drf-spectacular==0.28.0      # OpenAPI 3 schema / Swagger UI

# ── Dev / Test ────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
